cheri: Fix warnings in fs/smb/
Only minor unsigned long vs. uintptr_t issues.

+++ fs/smb/client/cifsglob.h
@@ -571,7 +571,7 @@ struct smb_version_operations {
 				struct cifs_tcon *tcon,
 				struct cifs_sb_info *cifs_sb,
 				__le16 *path, int is_dir,
-				unsigned long p);
+				uintptr_t p);
 	/* make unix special files (block, char, fifo, socket) */
 	int (*make_node)(unsigned int xid,
 			 struct inode *inode,
@@ -2205,7 +2205,7 @@ static inline int cifs_get_num_sgs(const struct smb_rqst *rqst,
 {
 	unsigned int len, skip;
 	unsigned int nents = 0;
-	unsigned long addr;
+	uintptr_t addr;
 	size_t data_size;
 	int i, j;
 
@@ -2242,7 +2242,7 @@ static inline int cifs_get_num_sgs(const struct smb_rqst *rqst,
 		for (j = 0; j < rqst[i].rq_nvec; j++) {
 			struct kvec *iov = &rqst[i].rq_iov[j];
 
-			addr = (unsigned long)iov->iov_base + skip;
+			addr = (uintptr_t)iov->iov_base + skip;
 			if (unlikely(is_vmalloc_addr((void *)addr))) {
 				len = iov->iov_len - skip;
 				nents += DIV_ROUND_UP(offset_in_page(addr) + len,
@@ -2266,7 +2266,7 @@ static inline void cifs_sg_set_buf(struct sg_table *sgtable,
 				   const void *buf,
 				   unsigned int buflen)
 {
-	unsigned long addr = (unsigned long)buf;
+	uintptr_t addr = (uintptr_t)buf;
 	unsigned int off = offset_in_page(addr);
 
 	addr &= PAGE_MASK;
